Getting docker tests to work on an amd64 mac
I was getting the sane error as this person trying to run the tests on
an M2 mac

https://stackoverflow.com/questions/71040681/qemu-x86-64-could-not-open-lib64-ld-linux-x86-64-so-2-no-such-file-or-direc

So I adopted the solution of the first answerer and then had some bugs
due to the the containers not installing curl or python. I opted for
python3 which meant changing the pip command too.