I’ve updated the project to test the Jinja macros against the fixtures generated as part of the GOV.UK Frontend package.
This removes the need to use and maintain the `govuk-frontend-diff` project.

I know in the `govuk-frontend-diff` project it’s set up to give quite a detailed diff if the generated HTML does not match and the way I’ve set it up, to compare using xmldiff, may not be quite what
you’re after. I’m happy for any ideas to improve this.